requirements.txt
flask
flask-cors
stockfish
google-generativeai
python-chess
gunicorn